reworked, deleted Class 'Car1'
